add navbar and table view

+++ contabilidad-front/src/components/Movements.jsx
@@ -1,12 +1,41 @@
 import React, { useEffect, useState } from "react";
 import axios from "axios";
+import {
+  Paper,
+  makeStyles,
+  TableBody,
+  TableRow,
+  TableCell,
+  TableContainer,
+} from "@material-ui/core";
 import Button from "@material-ui/core/Button";
-import { makeStyles } from "@material-ui/core/styles";
 import DeleteIcon from "@material-ui/icons/Delete";
+import Notification from "./Notification";
+import useTable from "./useTable";
+
+const useStyles = makeStyles((theme) => ({
+  pageContent: {
+    margin: theme.spacing(5),
+    padding: theme.spacing(3),
+  },
+}));
+
+const headCells=[
+    {id:"Movement", label:"Where the money goes"},
+    {id:"Amount", label:"How much money"},
+    {id: "Delete", label:"Delete"}
+]
 
 const Movements = () => {
   const [movements, setMovements] = useState([]);
   const [error, setError] = useState("");
+  const [notify, setNotify] = useState({
+    isOpen: false,
+    message: "",
+    type: "",
+  });
+  const classes = useStyles();
+  const { TblContainer, TblHead } = useTable(movements, headCells);
 
   function getMovements() {
     let options = {
@@ -31,17 +60,24 @@ const Movements = () => {
   }, []);
 
   async function deleteMovement(id) {
-    let options = {
-      method: "delete",
-      url: `http://localhost:3001/api/deletemovement/${id}`,
-      crossdomain: true,
-    };
-    try {
-      console.log("el id", id);
-      await axios(options);
-      getMovements();
-    } catch (err) {
-      console.log("ENTORAL ELES", err);
+    if (window.confirm("Are you sure to delete this record?")) {
+      let options = {
+        method: "delete",
+        url: `http://localhost:3001/api/deletemovement/${id}`,
+        crossdomain: true,
+      };
+      try {
+        console.log("el id", id);
+        await axios(options);
+        getMovements();
+        setNotify({
+          isOpen: true,
+          message: "Deleted successfully",
+          type: "success",
+        });
+      } catch (err) {
+        console.log("ENTORAL ELES", err);
+      }
     }
   }
 
@@ -52,44 +88,62 @@ const Movements = () => {
     whiteSpace: "nowrap",
   };
 
-  const useStyles = makeStyles((theme) => ({
-    button: {
-      margin: theme.spacing(1),
-    },
-  }));
-  const classes = useStyles();
-
   return (
-    <div>
-      <h1 id="mayuscula">Movimientos</h1>
-      <h2>{error}</h2>
-      <div className="card-deck product-list">
-        {movements.map((movement) => (
-          <div className="col-sm-3">
-            <div className="card mb-3" style={myStyle}>
-              <Button
-                onClick={() => deleteMovement(movement._id)}
-                variant="contained"
-                color="secondary"
-                style={{
-                  position: "absolute",
-                  right: 5,
-                  top: 5,
-                }}
-                className={classes.button}
-                startIcon={<DeleteIcon />}
-              ></Button>
-
-              <div className="card-body">
-                <h5 className="card-title">{movement.name}</h5>
-                <p className="card-text"> $ {movement.amount}</p>
-              </div>
-            </div>
-          </div>
-        ))}
-      </div>
-    </div>
+    <>
+      <Paper className={classes.pageContent}>
+        <TblContainer>
+            <TblHead />
+          <TableBody>
+            {movements.map((movement) => (
+              <TableRow key={movement._id}>
+                <TableCell>{movement.name}</TableCell>
+                <TableCell>${movement.amount}</TableCell>
+                <TableCell>
+                <Button
+          onClick={() => deleteMovement(movement._id)}
+          variant="contained"
+          color="secondary"
+          startIcon={<DeleteIcon />}
+        ></Button>
+                </TableCell>
+              </TableRow>
+            ))}
+          </TableBody>
+        </TblContainer>
+      </Paper>
+      <Notification notify={notify} setNotify={setNotify} />
+    </>
   );
 };
 
 export default Movements;
+
+{
+  /* <h1 id="mayuscula">Movimientos</h1>
+<h2>{error}</h2>
+<div className="card-deck product-list">
+  {movements.map((movement) => (
+    <div className="col-sm-3">
+      <div className="card mb-3" style={myStyle}>
+        <Button
+          onClick={() => deleteMovement(movement._id)}
+          variant="contained"
+          color="secondary"
+          style={{
+            position: "absolute",
+            right: 5,
+            top: 5,
+          }}
+          startIcon={<DeleteIcon />}
+        ></Button>
+
+        <div className="card-body">
+          <h5 className="card-title">{movement.name}</h5>
+          <p className="card-text"> $ {movement.amount}</p>
+        </div>
+      </div>
+    </div>
+  ))}
+</div>
+<Notification notify={notify} setNotify={setNotify} /> */
+}
